fix : undefined of rootBoundary.rootTarget
